Fix build errors and resolve ambiguous enum reference

+++ Server/src/Aetherfall.Domain/Characters/CharacterFormulaService.cs
@@ -46,7 +46,7 @@ namespace Aetherfall.Domain.Characters
                 MaxStamina: (decimal)CalculateStamina(level, attributes.Dexterity, attributes.Vitality),
                 PhysicalDamage: (decimal)physicalDamage,
                 SpellDamage: (decimal)spellDamage,
-                CriticalChance: criticalChance
+                CriticalChance: (decimal)criticalChance
             );
         }
     }
